requirements.txt
fastapi
uvicorn
streamlit
pandas
numpy
scikit-learn
xgboost
sentence-transformers
transformers
torch
faiss-cpu
joblib
langdetect
openai-whisper
moviepy
ffmpeg-python
librosa
soundfile
pydantic